Introduce `asakusafwOrganizer.{core,mapreduce}.version`.
