+++ gradle/src/main/groovy/com/asakusafw/mapreduce/gradle/plugins/internal/AsakusaMapReduceSdkBasePlugin.groovy
@@ -15,14 +15,9 @@
  */
 package com.asakusafw.mapreduce.gradle.plugins.internal
 
-import java.util.regex.Matcher
-import java.util.regex.Pattern
-
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
-import com.asakusafw.gradle.plugins.AsakusafwCompilerExtension
-import com.asakusafw.gradle.plugins.AsakusafwPluginConvention
 import com.asakusafw.gradle.plugins.AsakusafwSdkExtension
 import com.asakusafw.gradle.plugins.AsakusafwSdkPluginParticipant
 import com.asakusafw.gradle.plugins.internal.AsakusaSdkPlugin
@@ -30,16 +25,12 @@ import com.asakusafw.gradle.plugins.internal.PluginUtils
 
 /**
  * A base plug-in of {@link AsakusaMapReduceSdkPlugin}.
- * This only organizes conventions and dependencies.
+ * This only organizes dependencies and testkits.
  * @since 0.9.0
  */
 class AsakusaMapReduceSdkBasePlugin implements Plugin<Project> {
 
-    private static final Pattern OPTION_PATTERN = ~/([\+\-])\s*([0-9A-Za-z_\\-]+)|(X[0-9A-Za-z_\\-]+)=([^,]*)/
-
     private Project project
-
-    private AsakusafwCompilerExtension extension
 
     @Override
     void apply(Project project) {
@@ -47,24 +38,15 @@ class AsakusaMapReduceSdkBasePlugin implements Plugin<Project> {
 
         project.apply plugin: AsakusaSdkPlugin
         project.apply plugin: AsakusaMapReduceBasePlugin
-        this.extension = AsakusaSdkPlugin.get(project).extensions.create('mapreduce', AsakusafwCompilerExtension)
 
-        configureConvention()
+        configureTestkit()
         configureConfigurations()
     }
 
-    private void configureConvention() {
-        AsakusaMapReduceBaseExtension base = AsakusaMapReduceBasePlugin.get(project)
-        AsakusafwPluginConvention sdk = AsakusaSdkPlugin.get(project)
-        extension.conventionMapping.with {
-            outputDirectory = { project.file(sdk.compiler.compiledSourceDirectory) }
-            runtimeWorkingDirectory = { sdk.compiler.hadoopWorkDirectory }
-            compilerProperties = { parseOptions(sdk.compiler.compilerOptions) }
-            failOnError = { true }
-        }
-        PluginUtils.injectVersionProperty(extension, { base.featureVersion })
-        sdk.sdk.availableTestkits << new AsakusaMapReduceTestkit()
-        sdk.sdk.availableTestkits << new AsakusaSimpleMapReduceTestkit()
+    private void configureTestkit() {
+        AsakusafwSdkExtension sdk = AsakusaSdkPlugin.get(project).sdk
+        sdk.availableTestkits << new AsakusaMapReduceTestkit()
+        sdk.availableTestkits << new AsakusaSimpleMapReduceTestkit()
     }
 
     private void configureConfigurations() {
@@ -107,48 +89,6 @@ class AsakusaMapReduceSdkBasePlugin implements Plugin<Project> {
                 }
             }
         }
-    }
-
-    private Map<String, Object> parseOptions(List<String> options) {
-        Map<String, Object> results = [:]
-        for (String s : options) {
-            if (s == null || s.trim().isEmpty()) {
-                continue
-            }
-            String option = s.trim()
-            Matcher m = OPTION_PATTERN.matcher(option)
-            if (m.matches()) {
-                if (m.group(1) != null) {
-                    String key = m.group(2)
-                    boolean value = m.group(1) == '+'
-                    results[key] = value
-                } else if (m.group(3) != null) {
-                    String key = m.group(3)
-                    String value = m.group(4)
-                    results[key] = value
-                } else {
-                    throw new AssertionError(option)
-                }
-            } else {
-                project.logger.warn "unrecognizable compiler option: ${option}"
-            }
-        }
-        return results
-    }
-
-    /**
-     * Returns the extension object of this plug-in.
-     * The plug-in will be applied automatically.
-     * @param project the target project
-     * @return the related extension
-     */
-    static AsakusafwCompilerExtension get(Project project) {
-        project.apply plugin: AsakusaMapReduceSdkBasePlugin
-        AsakusaMapReduceSdkBasePlugin plugin = project.plugins.getPlugin(AsakusaMapReduceSdkBasePlugin)
-        if (plugin == null) {
-            throw new IllegalStateException('AsakusaMapReduceSdkBasePlugin has not been applied')
-        }
-        return plugin.extension
     }
 
     /**
